+++ CHANGELOG.md
@@ -36,6 +36,8 @@
 * Add mockery mocks for both `Customer` / `CustomerIdentityService` for easier testing
 * Add `State` field to customer address to be closer to cart address type, expose via GraphQL
 
+**price**
+* When marshalling `domain.Price` to JSON the amount is rounded.
 
 ## v3.4.0
 **cart**